Refactor `iter_store_keys`  (#2651)
## Description
<!-- List of detailed changes -->
Was looking into:
```rs
                // We have to clone the prefix and start, because we need to pass them to the iterator
                // if someone finds a solution without making it a vec, feel free to contribute :)
                let column = column.id();
                let prefix = prefix.map(|prefix| prefix.to_vec());
                let start = start.map(|start| start.to_vec());
```
and did some refactoring to get to the bottom of the problem.

I don't know if there is a solution unless we decide to change the
signature of `iter_store_keys` to something like:

```rs
    fn iter_store_keys<'a>(
        &'a self,
        column: Self::Column,
        prefix: Option<&'a [u8]>,
        start: Option<&'a [u8]>,
        direction: IterDirection,
    ) -> BoxedIter<KeyItem>;
```
But that's not realistic in this PR.

Might as well keep the refactoring because I feel like it's cleaner.