Add integration skeleton of parallel-executor (#2617)
## Description

This PR is the first one of a sequel of PR that will introduce the
`parallel-executor` to the codebase.

The goal of this first PR is to create the new module and showcase the
smooth integration with the existing codebase.

The instantiation of the `parallel-executor` has been commented to pass
the tests with `--all-features` however I want to keep them to showcase
what the integration looks like and will be uncommented when real code
exists.

## Checklist
- [x] Breaking changes are clearly marked as such in the PR description
and changelog
- [x] New behavior is reflected in tests
- [x] [The specification](https://github.com/FuelLabs/fuel-specs/)
matches the implemented behavior (link update PR if changes are needed)

### Before requesting review
- [x] I have reviewed the code myself
- [x] I have created follow-up issues caused by this PR and linked them
here