Add versions on readme wiht action to update (#2647)
## Description
I added the versions used in different networks for other people to
easily have the information. To avoid technical debt and maintenance I
added an Action that chekcs every hour if a network has been updated and
creates a PR to update the table if needed.

This has been tested on a fork of chain-configuration repository.

Job that create PR :
https://github.com/AurelienFT/chain-configuration/actions/runs/13032692218/job/36355422292
PR : https://github.com/AurelienFT/chain-configuration/pull/6
Job when a PR already exists :
https://github.com/AurelienFT/chain-configuration/actions/runs/13032715720/job/36355503047
Job when the versions are up-to-date :
https://github.com/AurelienFT/chain-configuration/actions/runs/13032730000/job/36355548982

## Checklist
- [x] Breaking changes are clearly marked as such in the PR description
and changelog
- [x] New behavior is reflected in tests
- [x] [The specification](https://github.com/FuelLabs/fuel-specs/)
matches the implemented behavior (link update PR if changes are needed)

### Before requesting review
- [x] I have reviewed the code myself
- [x] I have created follow-up issues caused by this PR and linked them
here


Note: We may have a problem with CLA such as we have with cargo update
and so I think we need to configure CLA to allow github action user.