Increase the P2P message limits (#2705)
## Description

Increase the P2P limits to allow to sync correctly. Previously the
default number of blocks in a batch could have been higher than the
default maximum size. Now the default values for each parameters are
aligned.

(We discovered it with big blocks on various networks)

## Checklist
- [x] Breaking changes are clearly marked as such in the PR description
and changelog
- [x] New behavior is reflected in tests
- [x] [The specification](https://github.com/FuelLabs/fuel-specs/)
matches the implemented behavior (link update PR if changes are needed)

### Before requesting review
- [x] I have reviewed the code myself
- [x] I have created follow-up issues caused by this PR and linked them
here