Remove reviewers in update README action (#2706)
## Description

Currently, all the action readme update jobs are failing because the
group "FuelLabs/client" can't be seen by the Action and so it can't
assign it as reviewer. This PR comes back to what we do with cargo
update, we just let our repository rules determine the reviewers.

## Checklist
- [x] Breaking changes are clearly marked as such in the PR description
and changelog
- [x] New behavior is reflected in tests
- [x] [The specification](https://github.com/FuelLabs/fuel-specs/)
matches the implemented behavior (link update PR if changes are needed)

### Before requesting review
- [x] I have reviewed the code myself
- [x] I have created follow-up issues caused by this PR and linked them
here