New txpool worker to remove lock contention (#2725)
## Description

### Problem

When running benches on transaction pool we observed that each thread
trying to submit to the same pool with a thread was a costly behavior.

### Solution in this PR

We changed the pool to live only in one thread and all communications
are managed through channels where priority can be explicitly defined
between each tasks.

### Side modifications

Now we are sending tx through P2P after their full insertion inside the
pool. Before it was done after one round of verification but before the
final checks and the pool integration.

Now we are not doing any pool verifications before computing the
predicates and the signatures. This pool and input verifications step is
now done after the predicate and signatures verifications inside of the
pool worker thread.

## Checklist
- [x] Breaking changes are clearly marked as such in the PR description
and changelog
- [x] New behavior is reflected in tests
- [x] [The specification](https://github.com/FuelLabs/fuel-specs/)
matches the implemented behavior (link update PR if changes are needed)

### Before requesting review
- [x] I have reviewed the code myself
- [x] I have created follow-up issues caused by this PR and linked them
here

---------

Co-authored-by: green <xgreenx9999@gmail.com>