Dedicated threads for the block importer (#2750)
Removed cloning of the block and utilized the feature of rayon runtime
with scoped parallel jobs execution.
Moved code to catch `InvalidDatabaseStateAfterExecution` to the right
place.
Removed requirements from the database to be modifiable. 
Fixed the not working `async` part of the block importer. Before
importing of the block were blocking the tokio task and `await` never
was reached.