Add possibility to submit a list of changes to rocksdb (#2619)
## Description

The goal of this PR is to allow to insert a list of changes instead of
one `HashMap` of changes. This avoid multiple update produced by
different threads for example, to be merged in one transaction.

By removing this step we remove a whole iteration and insert/deletion in
a big `HashMap` of changes.

It's currently not used anywhere in the code but you can see a usage
example here : https://github.com/FuelLabs/fuel-core/pull/2618

In this PR when we have to write history modifications we are still
merging the changes in one big transaction before calling
`store_modifications_history` because I don't think it's possible to use
this algorithm with changes splitted in different parts. @acerone85 if
you have an idea I would be happy to here it but in any case I think
this should go to follow-up PR to not overcharge this one. (I made an
attempt in branch [try_adapt_history_vec
branch](https://github.com/FuelLabs/fuel-core/tree/try_adapt_history_vec)
but it's mehhh)

## Checklist
- [x] Breaking changes are clearly marked as such in the PR description
and changelog
- [x] New behavior is reflected in tests
- [x] [The specification](https://github.com/FuelLabs/fuel-specs/)
matches the implemented behavior (link update PR if changes are needed)

### Before requesting review
- [x] I have reviewed the code myself
- [x] I have created follow-up issues caused by this PR and linked them
here

---------

Co-authored-by: Mitchell Turner <james.mitchell.turner@gmail.com>
Co-authored-by: green <xgreenx9999@gmail.com>