GraphQL: historical contract storage and balances (#2682)
~~Requires https://github.com/FuelLabs/fuel-core/pull/2661~~

Closes https://github.com/FuelLabs/fuel-core/issues/857

## Description

Adds GraphQL APIs for querying contract storage slots and balances at
given block height. This is behind `--historical-execution` flag, as it
can be quite expensive.

This PR also fixes the issue with RPC consistency feature for the
subscriptions(without the fix first we perform the logic of the query,
and only after verify the required height).

## Checklist
- [x] Breaking changes are clearly marked as such in the PR description
and changelog
- [x] New behavior is reflected in tests

### Before requesting review
- [x] I have reviewed the code myself
- [x] I have created follow-up issues caused by this PR and linked them
here

---------

Co-authored-by: green <xgreenx9999@gmail.com>
Co-authored-by: AurelienFT <32803821+AurelienFT@users.noreply.github.com>