Deprecate dryRun mutation, add query impl for it (#2781)
Closes #2717

## Description

For some reason, dryRun was originally created as a GraphQL mutation
instead of a query. Since the operation doesn't actually perform any
modifications, this seems like a mistake. This PR adds identical query,
and marks the mutation is deprecated. No changes to the client are done,
so it continues working with older versions. This is completely
non-breaking change, but prepares us to potentially perform the breaking
change in the future.

## Checklist
- [ ] Breaking changes are clearly marked as such in the PR description
and changelog
- [ ] New behavior is reflected in tests
- [ ] [The specification](https://github.com/FuelLabs/fuel-specs/)
matches the implemented behavior (link update PR if changes are needed)

### Before requesting review
- [ ] I have reviewed the code myself
- [ ] I have created follow-up issues caused by this PR and linked them
here