Add allow_partial parameter to the "coins to spend" query (#2912)
## Linked Issues/PRs
Closes https://github.com/FuelLabs/fuel-core/issues/2898

THIS IS A REVIVAL OF https://github.com/FuelLabs/fuel-core/pull/2904
WITH REVIEW COMMENTS ADDRESSED (because of CLA). Inspiration from Rafal.
Kudos to @rafal-ch !

## Description
This PR introduces the new allow_partial parameter to the "coins to
spend" inputs query. This field is optional to avoid breaking change. If
set to true, the query returns available coins instead of failing when
the requested amount is unavailable.

I consider this change "non-breaking" since the new parameter is not
strictly required when sending the query.

## Checklist
- [x] New behavior is reflected in tests

### Before requesting review
- [x] I have reviewed the code myself

---------

Co-authored-by: Mitchell Turner <james.mitchell.turner@gmail.com>