Make preconfirmation optional on API endpoints (#2925)
## Linked Issues/PRs
Closes https://github.com/FuelLabs/fuel-core/issues/2924

## Description

| Endpoint affected | Changes realised | Tested where? | 
|-|-|-|
| statusChange | Filter the preconfirmations depending on the boolean
values after conversion of the transaction status to the API one. | In
preconfirmation integration tests |
| submitAndAwaitStatus | Filter the preconfirmations depending on the
boolean values after conversion of the transaction status to the API
one. | In preconfirmation integration tests |
| `status` field of `Transaction` object | Filter the preconfirmations
depending on the boolean values after conversion of the transaction
status to the API one. For preconfirmation we still want to send a
status that will be Submitted. | Client doesn't support this feature so
no explicit tests |

## Checklist
- [x] Breaking changes are clearly marked as such in the PR description
and changelog
- [x] New behavior is reflected in tests
- [x] [The specification](https://github.com/FuelLabs/fuel-specs/)
matches the implemented behavior (link update PR if changes are needed)

### Before requesting review
- [x] I have reviewed the code myself
- [x] I have created follow-up issues caused by this PR and linked them
here

---------

Co-authored-by: Aaryamann Challani <43716372+rymnc@users.noreply.github.com>
Co-authored-by: green <xgreenx9999@gmail.com>