chore: Additional error logs for eth provider failures (#3004)
This PR adds additional error logs when we have Ethereum provider
failures. We need this specifically because the upstream `QuorumError`
doesn't include any provider errors in the `NoQuorumReached` variant:

```
#[derive(Error, Debug)]
/// Error thrown when sending an HTTP request
pub enum QuorumError {
    #[error("No Quorum reached.")]
    /// NoQuorumReached
    NoQuorumReached {
        /// Returned responses
        values: Vec<Value>,
        /// Returned errors
        errors: Vec<ProviderError>,
    },
}
```

However, by adding logs of the debug representation of this error we'll
expose this information which can be helpful when debugging quorum
failures.