Fix client coins endpoint to return all asset types (#2984)
## Linked Issues/PRs
Closes #2960 

<!-- List of related issues/PRs -->
## Description
<!-- List of detailed changes -->
This PR fixes the behavior of the `coins` endpoint on the client side.
Previously, if `asset_id` was `None`, it defaulted to
`AssetId::default()`, which is incorrect. The expected behavior (and
what the query side already supports) is to return all assets when
`asset_id` is `None`.
## Checklist
- [ ] Breaking changes are clearly marked as such in the PR description
and changelog
- [ ] New behavior is reflected in tests
- [ ] [The specification](https://github.com/FuelLabs/fuel-specs/)
matches the implemented behavior (link update PR if changes are needed)

### Before requesting review
- [✓] I have reviewed the code myself
- [ ] I have created follow-up issues caused by this PR and linked them
here

### After merging, notify other teams

[Add or remove entries as needed]

- [ ] [Rust SDK](https://github.com/FuelLabs/fuels-rs/)
- [ ] [Sway compiler](https://github.com/FuelLabs/sway/)
- [ ] [Platform
documentation](https://github.com/FuelLabs/devrel-requests/issues/new?assignees=&labels=new+request&projects=&template=NEW-REQUEST.yml&title=%5BRequest%5D%3A+)
(for out-of-organization contributors, the person merging the PR will do
this)
- [ ] Someone else?

---------

Co-authored-by: Green Baneling <XgreenX9999@gmail.com>