Add test showing mint added (#3061)
## Linked Issues/PRs
closes https://github.com/FuelLabs/fuel-core/issues/1867

## Description
The issue says we should have a test showing that incorrect mint values
won't get validated, but that's not how the code works--expecially from
the external apis. There is no way for us to produce a failing tests for
either the validate or the produce cases. validate uses the provided
mint to choose a value of gas_price and execute creates the mint using
whatever gas_price you provide.

Instead, this just adds a test showing that a mint is added when you
call the external api `produce_without_commit_with_source`. There is
already a test for `validate` (technically `validate_and_commit` which
is a bastardization of `validate` only used in tests).

The test coverage for block production and validation is very bad right
now, as captured in this issue:
https://linear.app/fuel-network/issue/CLNT-1645/beef-up-execution-tests

We should be very clear about what external apis each test is for,
instead many of them are for abstract concepts.

### Before requesting review
- [x] I have reviewed the code myself

---------

Co-authored-by: Hannes Karppila <2204863+Dentosal@users.noreply.github.com>