feat: Support private keys in coin config and log these if present (#3014)
This PR extends the `CoinConfig` in the state configuration to hold
either an owner address or a private key. Additionally, I've added info
logs when the node is started from a local testnet reader to print
these.

The purpose is to make it easier to find and keep track of secret keys
with funds when running a local node.

---------

Co-authored-by: Green Baneling <XgreenX9999@gmail.com>