Block Aggregator Service Scaffolding (#3085)
## Linked Issues/PRs
<!-- List of related issues/PRs -->
Starts https://github.com/FuelLabs/fuel-core/issues/3086

Follow-up Issues:
https://github.com/FuelLabs/fuel-core/issues/3087
https://github.com/FuelLabs/fuel-core/issues/3088
https://github.com/FuelLabs/fuel-core/issues/3089
https://github.com/FuelLabs/fuel-core/issues/3090
https://github.com/FuelLabs/fuel-core/issues/3091

## Description
<!-- List of detailed changes -->
This PR adds the core logic of the block aggregator service. The domain
includes:
1. Receiving the block from some abstracted source. This will be
implemented later as a combination of block importer and access to the a
on-chain db, where it will sync any missing blocks before subscribing to
the importer values
2. Storing the blocks as raw bytes (? maybe we can abstract the
serialization format). This will be implemented as a normal Fuel DB
3. An API abstraction with queries for ranges of blocks as well as info
on the current height of the aggregator's db. This could be implemented
as some protobuf server or something else.

We don't want this service to block the initialization of other
services. This means that syncing will happen in the `run` function
(rather than in some `UninitializedTask`). As the syncing progresses,
the users can query the current height with the
`BlockAggregatorQuery::GetCurrentHeight`.

The response type for a block range query is a assoc type because we are
still up in the air about how it will be handled. There is a kinda of
example type added, but it's not used anywhere yet, including the tests:
```rs
/// The response to a block range query, either as a literal stream of blocks or as a remote URL
pub enum BlockRangeResponse {
    /// A literal stream of blocks
    Literal(BoxStream<Block>),
    /// A remote URL where the blocks can be fetched
    Remote(String),
}
```

## Checklist
- [x] New behavior is reflected in tests

### Before requesting review
- [x] I have reviewed the code myself
- [x] I have created follow-up issues caused by this PR and linked them
here

---------

Co-authored-by: Aaryamann Challani <43716372+rymnc@users.noreply.github.com>