Chore/parallel executor benches (#3045)
## Linked Issues/PRs
<!-- List of related issues/PRs -->

## Description
<!-- List of detailed changes -->
This in practice ended up being as much of an integration tests as a
benchmark. There were a few places that weren't plugged in that needed
finessing before this would run properly.

~In addition, this uses the `u32-tx-pointer` feature from `fuel-tx`.
This is a **_BREAKING_** change as the header has changed.~

~If we don't want to contaminate the code with a breaking change for the
benchmark, then we can hide that behind a new feature flag on the
`fuel-types` library.~

I've introduced a `u32-tx-count` feature that changes a lot of `u16` to
`u32` for tx pointers as well as tx count in header etc. This prevents
this code from being breaking.

## Checklist
- [ ] New behavior is reflected in tests
- [ ] [The specification](https://github.com/FuelLabs/fuel-specs/)
matches the implemented behavior (link update PR if changes are needed)

### Before requesting review
- [x] I have reviewed the code myself
- [ ] I have created follow-up issues caused by this PR and linked them
here