Removed SMT usage from STF and benchmarks (#3117)
In https://github.com/FuelLabs/fuel-core/pull/3099 we removed usage of
SMT from STF. This PR, cleans up remaining code.

The columns in the database still must exist, for backward
compatibility. But outside of that code can be cleaned up.

The algorithm to calculate the genesis state root and the balance root
was changed to not use SMTs. This change doesn't affect mainnet,
testnet, and devnet, because these networks didn't use any balances or
storage slots in the state config. But the change is breaking for anyone
who used them.

---------

Co-authored-by: Hannes Karppila <hannes.karppila@gmail.com>